pyproject.toml
[project]
name = "greptimedb-sqlalchemy"
version = "0.1.3"
description = "SQLAlchemy support for GreptimeDB"
authors = [
    { name = "Ning Sun", email = "sunning@greptime.com" }
]
dependencies = [
    "sqlalchemy>=1.4, <2",
    "psycopg2-binary>=2.9.9",
]
readme = "README.md"
requires-python = ">= 3.9"
classifiers = [
    'Intended Audience :: Developers',
    'Topic :: Database',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]

[project.entry-points.'sqlalchemy.dialects']
greptimedb = 'greptimedb_sqlalchemy:GreptimeDBDialect'

[project.entry-points.'superset.db_engine_specs']
greptimedb = 'greptimedb.superset:GreptimeDBEngineSpec'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ['apache-superset>=3.0.0']

[tool.hatch.metadata]
allow-direct-references = true
